Refactor Bicep parser to replace IgnoreMap with bicepComment for improved comment handling and remove unused IgnoreMap file
